+++ examples/switch_systematic_on_off/switch_systematic_on_off.cpp
@@ -30,15 +30,12 @@ int main(void)
     uint32_t max_symbols = 10;
     uint32_t max_symbol_size = 100;
 
-    bool trace_enabled = true;
-
     // Initilization of encoder and decoder
     kodocpp::encoder_factory encoder_factory(
         kodo_full_vector,
         kodo_binary8,
         max_symbols,
-        max_symbol_size,
-        trace_enabled);
+        max_symbol_size);
 
     kodocpp::encoder encoder = encoder_factory.build();
 
@@ -46,8 +43,7 @@ int main(void)
         kodo_full_vector,
         kodo_binary8,
         max_symbols,
-        max_symbol_size,
-        trace_enabled);
+        max_symbol_size);
 
     kodocpp::decoder decoder = decoder_factory.build();
 
@@ -63,7 +59,11 @@ int main(void)
     // Just for fun - fill the data with random data
     std::generate(data_in.begin(), data_in.end(), rand);
 
-    encoder.set_symbols(data_in.data(), encoder.block_size());
+    encoder.set_const_symbols(data_in.data(), encoder.block_size());
+
+    // Set the storage for the decoder
+    std::vector<uint8_t> data_out(decoder.block_size());
+    decoder.set_mutable_symbols(data_out.data(), decoder.block_size());
 
     std::cout << "Start encoding / decoding" << std::endl;
     while (!decoder.is_complete())
@@ -108,9 +108,6 @@ int main(void)
         std::cout << "Symbols decoded: " << decoder.symbols_uncoded()
                   << std::endl;
     }
-
-    std::vector<uint8_t> data_out(decoder.block_size());
-    decoder.copy_from_symbols(data_out.data(), decoder.block_size());
 
     // Check if we properly decoded the data
     if (std::equal(data_out.begin(), data_out.end(), data_in.begin()))