+++ test/src/test_full_vector_codes.cpp
@@ -9,17 +9,25 @@
 
 #include "test_helper.hpp"
 
-TEST(test_full_vector_codes, invoke_api)
+
+TEST(test_full_vector_codes, basic_api)
 {
     using namespace kodocpp;
 
     uint32_t max_symbols = rand_symbols();
     uint32_t max_symbol_size = rand_symbol_size();
 
-    kodo_code_type code_type = kodo_full_vector;
+    test_basic_api(kodo_full_vector, max_symbols, max_symbol_size);
+}
 
-    test_basic_api(max_symbols, max_symbol_size, code_type, kodo_binary);
-    test_basic_api(max_symbols, max_symbol_size, code_type, kodo_binary4);
-    test_basic_api(max_symbols, max_symbol_size, code_type, kodo_binary8);
-    test_basic_api(max_symbols, max_symbol_size, code_type, kodo_binary16);
+TEST(test_full_vector_codes, sparse_api)
+{
+    using namespace kodocpp;
+
+    // Make sure that the decoding can complete with a lower density
+    uint32_t max_symbols = rand_symbols() + 10;
+    uint32_t max_symbol_size = rand_symbol_size();
+
+    test_basic_api(kodo_sparse_full_vector, kodo_full_vector, max_symbols,
+                   max_symbol_size);
 }