+++ test/src/test_sparse_seed_codes.cpp
@@ -9,12 +9,13 @@
 
 #include "test_helper.hpp"
 
-TEST(test_sliding_window_codes, invoke_api)
+TEST(test_sparse_seed_codes, basic_api)
 {
     using namespace kodocpp;
 
-    uint32_t max_symbols = rand_symbols();
+    // Make sure that the decoding can complete with a lower density
+    uint32_t max_symbols = rand_symbols() + 10;
     uint32_t max_symbol_size = rand_symbol_size();
 
-    test_basic_api(kodo_sliding_window, max_symbols, max_symbol_size);
+    test_basic_api(kodo_sparse_seed, max_symbols, max_symbol_size);
 }